.azd-setup/src/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.29.0
gunicorn==23.0.0
azure-identity==1.19.0
aiohttp==3.11.1

azure_ai_agents==1.0.0
azure_ai_projects==1.0.0b11

azure-core==1.34.0  # other versions might not compatible
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry==1.6.9 # version such as 1.6.11 isn't compatible
azure-search-documents
opentelemetry-sdk
setuptools==80.9.0
starlette>=0.40.0 # fix vulnerability
jinja2 # new dependent of fastapi